use Handler instead of knHandler
