+++ swagger/models.json
@@ -166,16 +166,23 @@
       "description": "BaseImage base image",
       "type": "object",
       "required": [
+        "name",
         "dockerUrl",
-        "language",
-        "name"
+        "language"
       ],
       "properties": {
+        "backingObject": {
+          "description": "BackingObject",
+          "type": "object",
+          "x-go-name": "BackingObject",
+          "readOnly": true
+        },
         "createdTime": {
-          "description": "created time",
+          "description": "CreatedTime",
           "type": "integer",
           "format": "int64",
-          "x-go-name": "CreatedTime"
+          "x-go-name": "CreatedTime",
+          "readOnly": true
         },
         "dockerUrl": {
           "description": "docker Url",
@@ -191,13 +198,13 @@
           "x-go-name": "Groups"
         },
         "id": {
-          "description": "id",
+          "description": "ID",
           "type": "string",
           "format": "uuid",
           "x-go-name": "ID"
         },
         "kind": {
-          "description": "kind",
+          "description": "Kind",
           "type": "string",
           "pattern": "^[\\w\\d\\-]+$",
           "x-go-name": "Kind",
@@ -208,14 +215,32 @@
           "type": "string",
           "x-go-name": "Language"
         },
-        "meta": {
-          "$ref": "#/definitions/Meta"
+        "modifiedTime": {
+          "description": "ModifiedTime",
+          "type": "integer",
+          "format": "int64",
+          "x-go-name": "ModifiedTime",
+          "readOnly": true
         },
         "name": {
-          "description": "name",
+          "description": "Name",
           "type": "string",
-          "pattern": "^[\\w\\d][\\w\\d\\-]*$",
+          "pattern": "^[\\w\\d][\\w\\d\\-]*[\\w\\d]|[\\w\\d]+$",
           "x-go-name": "Name"
+        },
+        "org": {
+          "description": "Org",
+          "type": "string",
+          "default": "default",
+          "pattern": "^[\\w\\d][\\w\\d\\-]*[\\w\\d]|[\\w\\d]+$",
+          "x-go-name": "Org"
+        },
+        "project": {
+          "description": "Project",
+          "type": "string",
+          "default": "default",
+          "pattern": "^[\\w\\d][\\w\\d\\-]*[\\w\\d]|[\\w\\d]+$",
+          "x-go-name": "Project"
         },
         "reason": {
           "description": "reason",
@@ -232,7 +257,7 @@
           "$ref": "#/definitions/Status"
         },
         "tags": {
-          "description": "tags",
+          "description": "Tags",
           "type": "array",
           "items": {
             "$ref": "#/definitions/Tag"
